Break apart ironflow nodes
